pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "psycopg2-binary>=2.9.10",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "streamlit>=1.49.1",
]
